add-updated-form

+++ AccessDataLayer/Entities/Product.cs
@@ -25,7 +25,7 @@ public class Product
     public ICollection<Order> Orders { get; set; } = new HashSet<Order>();
 
     [ForeignKey ("Supplier")]
-    public int SupId { get; set; }
+    public long SupId { get; set; }
     public Supplier Supplier { get; set; } = null!;
